worked today from 12 -22:40
set up the game screen with drawView and set up sql database. tried to
store all the survivors on the same database, used serializable to try
and store. had to change this and made a separate table to store the
survivors, appears to work not fully implemented. also talked to  my
brother in regards to the art got designs for the map should be able to
change snoopy to something more suited to the game