got some basic art from  my brother for the grid have display of the grid work up. data base functionality appears to be working, can store s
23rd Jan:
12 - 15 got drawings into the game and drawing of the grid onto the
screen
16:30 - 21:00 vhanged the text view on top of game activity to buttons
due to formatting error, database work numeros errors

24th - 25th JAN
18 - 21 after uni attempted to fix errors within the code still minor
errors

26th
12 - 18 fix database runn test appears to work as needed, cahnged
database structure

2nd Feb
10-13
methods for moving survivors between Activity nd nested draw view
14 -22
2 list created correctly transfer of info

3rd Feb
12 -15 save button created to save the current state and overrite where
neccessary
17 - 18 start of click method for the draw view

4th FEB
18 - 21 test save button

5th Feb
18:30 - 21 work on the click method to tell what sur are on the grid
that was clicked

6th Feb
11 -14 ixed matrix problem used grid draw collected the names of
survivors on the spot that has been clicked
15 - 18, 19 -20 attempt to display the list of survivors within a
dialog, not working needs to be fixed

7th FEB
18:30 - 20 trying to fix the array adapter on the dialog method it is
being created as null

14th FEB
15 - 22 changed the method of display now works as intended, found error
in the code stored to many values apon save no checking  for values
before adding them to the datatbase , grows to large to many entries
repeats. setup bones for the display of the info of the chosen survivor
from the list of survivors.