Feb 17th : 10 - 20(break from 17:45 -18:45) moving prob systems to the new grid game,  finished the system for displaying used survivors,
handleing safe points and farms,

need to add saving or which survivors have been used and what points are
safe, farms. it will require database work and altercations to the save
and load methods.

need to change the click method in draw view to consider where the
survivor is to alter what options are displayed to the player.